function that returns the natural square root of a number.

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -12,8 +12,8 @@
 
 int square_root_shadow(int num, int min, int max)
 {
-	int guess;
-	int guess_square;
+	int long guess;
+	int long guess_square;
 
 	if (min > max)
 		return (-1);
